practice_cases/simple_rag_assistant/requirements.txt
streamlit==1.46.0
langchain==0.3.26
langchain-chroma==0.2.4
langchain-community==0.3.27
langchain-core==0.3.79
langchain-deepseek==0.1.3
langchain-openai==0.3.19
langchain-huggingface==0.3.1
python-dotenv==1.1.0
pypdf==5.6.1
dashscope==1.23.5
tenacity==9.1.2
transformers==4.57.1 # 本地加载嵌入模型依赖
sentence-transformers==5.1.2  # 本地加载嵌入模型依赖
torch==2.9.0  # 本地加载嵌入模型依赖